Examples of learning under different priors.
With a wide prior, the decision-maker acts optimally to squash the outbreak early on. The decision-maker never really learns for sure if this was the right decision (i.e. if this was really necessary or if the outbreak would have stopped at low level), it is basically a precautionary reaction -- the posterior belief remains broad since learning opportunity is lost by crushing the outbreak.

In contrast, the decision-maker with a prior more confident in a low-level stable state (i.e. mistaking past evidence of the "ghost" for a stable state), fails to act decisively early on.  The posterior belief shows the decision-maker learns of this mistake, convering to a tight distribution on the true model but still stuck managing the outbreak continuously for ever.